split image filter into trait
